+++ src/BreakpointDriver.hpp
@@ -10,28 +10,22 @@ class BreakpointDriver
 public:
     explicit BreakpointDriver(void* process);
 
-    void set_base_address(InstructionPointer base_address);
+    void set_breakpoint(VirtualAddress va);
 
-    VirtualAddress ip_to_va(InstructionPointer ip) const;
-    InstructionPointer va_to_ip(VirtualAddress va) const;
+	[[nodiscard]] bool has_breakpoint(VirtualAddress va) const;
 
-    void set_breakpoint(InstructionPointer ip);
-
-	[[nodiscard]] bool has_breakpoint(InstructionPointer ip) const;
-
-    void remove_breakpoint(InstructionPointer ip);
+    void remove_breakpoint(VirtualAddress va);
 
 private:
     void* m_process;
-    InstructionPointer m_base_address;
 
     struct Breakpoint
     {
         std::byte original_instruction;
     };
 
-    std::unordered_map<InstructionPointer, Breakpoint> m_breakpoints;
+    std::unordered_map<VirtualAddress, Breakpoint> m_breakpoints;
 
-    void write_byte(InstructionPointer ip, std::byte byte);
+    void write_byte(VirtualAddress va, std::byte byte);
 };
 }