+++ src/types.hpp
@@ -8,16 +8,11 @@ namespace coverpp
 struct VirtualAddress
 {
     std::uintptr_t value;
-};
-
-struct InstructionPointer
-{
-    std::uintptr_t value;
 
     void* vp() const
     { return reinterpret_cast<void*>(value); }
 
-    bool operator==(const InstructionPointer&) const = default;
+    bool operator==(const VirtualAddress&) const = default;
 };
 }
 
@@ -27,28 +22,16 @@ struct std::formatter<coverpp::VirtualAddress>
     constexpr auto parse(std::format_parse_context& ctx)
     { return ctx.begin(); }
 
-    auto format(coverpp::VirtualAddress va, std::format_context& ctx) const
-    {
-        return std::format_to(ctx.out(), "0x{:X}", va.value);
-    }
-};
-
-template<>
-struct std::formatter<coverpp::InstructionPointer>
-{
-    constexpr auto parse(std::format_parse_context& ctx)
-    { return ctx.begin(); }
-
-    auto format(coverpp::InstructionPointer ip, std::format_context& ctx) const
+    auto format(coverpp::VirtualAddress ip, std::format_context& ctx) const
     {
         return std::format_to(ctx.out(), "0x{:X}", ip.value);
     }
 };
 
 template<>
-struct std::hash<coverpp::InstructionPointer>
+struct std::hash<coverpp::VirtualAddress>
 {
-    std::size_t operator()(coverpp::InstructionPointer ip) const
+    std::size_t operator()(coverpp::VirtualAddress ip) const
     {
         return std::hash<decltype(ip.value)>()(ip.value);
     }