+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This changelog follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.
 
 - Support more entrypoints (and make finding it optional): `wmain`, `WinMain`, `wWinMain`
 
+### Fixed
+
+- Initialize DIA session's base address
+- Proper use of virtual addresses throughout
+
 ## [26] - 2026-08-29
 
 Initial release.